fix: cast EXTRACT result to float8 for pgrx f64 compatibility
PostgreSQL 14+ EXTRACT() returns numeric (OID 1700). pgrx Datum
deserialization into Rust f64 expects float8 (OID 701); without the
explicit cast the pgrx SPI call fails with:
  Datum error: Postgres type numeric (Oid(1700)) is not compatible
  with the Rust type f64 (Oid(701))

Add ::float8 cast on the EXTRACT expression.

+++ pg-tide-ext/src/outbox.rs
@@ -275,7 +275,7 @@ fn outbox_status_impl(name: &str) -> Result<pgrx::JsonB, PgTideError> {
             "SELECT \
                COUNT(m.id) FILTER (WHERE m.consumed_at IS NULL)::bigint, \
                COUNT(m.id)::bigint, \
-               EXTRACT(epoch FROM now() - MIN(m.created_at) FILTER (WHERE m.consumed_at IS NULL)), \
+               EXTRACT(epoch FROM now() - MIN(m.created_at) FILTER (WHERE m.consumed_at IS NULL))::float8, \
                COALESCE(c.retention_hours, 24)::int \
              FROM tide.tide_outbox_config c \
              LEFT JOIN tide.tide_outbox_messages m ON m.outbox_name = c.outbox_name \
